BJTN DEVELOPMENT

+++ BJTN DEVELOPMENT/src/java/edu/uca/aca2016/impulse/web/ClientController.java
@@ -107,7 +107,7 @@ public class ClientController {
         context.put("list", list);
 
         int count = dao.getClientCount();
-        context.put("pages", Math.ceil((float) count / (float) total));
+        context.put("pages", Math.ceil(count / (float) total));
 
         context.put("page", pageid);
